+++ src/BigCommerce/Settings/Sections/Onboarding_Import_Settings.php
@@ -2,13 +2,13 @@
 
 namespace BigCommerce\Settings\Sections;
 
-
 use BigCommerce\Settings\Screens\Connect_Channel_Screen;
 
 class Onboarding_Import_Settings extends Settings_Section {
+
 	use Webhooks;
 
-	const NAME         = 'import_settings';
+	const NAME = 'import_settings';
 
 	public function register_settings_section() {
 		add_settings_section(
@@ -19,16 +19,23 @@ class Onboarding_Import_Settings extends Settings_Section {
 		);
 
 		add_settings_field(
-			Import::ENABLE_WEBHOOKS,
-			__( 'Enable Webhooks', 'bigcommerce' ),
-			[ $this, 'enable_webhooks_toggle' ],
+			Import::ENABLE_PRODUCTS_WEBHOOKS,
+			__( 'Enable Products Webhooks', 'bigcommerce' ),
+			[ $this, 'enable_products_webhooks_toggle' ],
 			Connect_Channel_Screen::NAME,
 			self::NAME
 		);
 
-		register_setting(
+		register_setting( Connect_Channel_Screen::NAME, Import::ENABLE_PRODUCTS_WEBHOOKS );
+
+		add_settings_field(
+			Import::ENABLE_CUSTOMER_WEBHOOKS,
+			__( 'Enable Customers Webhooks', 'bigcommerce' ),
+			[ $this, 'enable_customer_webhooks_toggle' ],
 			Connect_Channel_Screen::NAME,
-			Import::ENABLE_WEBHOOKS
+			self::NAME
 		);
+
+		register_setting( Connect_Channel_Screen::NAME, Import::ENABLE_CUSTOMER_WEBHOOKS );
 	}
 }